requirements.txt
# Mauritius Land Cover Classification
# Python 3.10+ recommended

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3

# Geospatial
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0
fiona>=1.9.0

# Google Earth Engine
earthengine-api>=0.1.350
geemap>=0.22.0

# Image Processing
opencv-python>=4.7.0
Pillow>=9.5.0
scikit-image>=0.21.0
albumentations>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web Interface
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Configuration & Logging
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Optional: Weights & Biases
# wandb>=0.15.0

# Optional: Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0

# GDAL (install via conda recommended)
# conda install -c conda-forge gdal

# Note: For GPU support, install PyTorch with CUDA:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
